make countdown non-ssr
