+++ package.json
@@ -62,6 +62,7 @@
     "lodash": "^4.17.11",
     "lodash.debounce": "^4.0.8",
     "lru-cache": "^4.1.3",
+    "luxon": "^1.9.0",
     "mdi-react": "4.4.0",
     "next": "^7.0.2-canary.49",
     "next-router-events": "^2.1.0",
